pyproject.toml
[project]
name = "schoology-ics"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "datetime~=5.5",
    "flask~=3.1.0",
    "icalendar~=6.3.1",
    "loguru~=0.7.3",
    "python-dotenv~=1.0.0",
    "requests~=2.32.3",
    "requests-oauthlib~=2.0.0",
    "schoolopy~=0.2.2",
    "setproctitle~=1.3.6",
]
